refactor(site-header): specs file, test and story adapted - TWIG-151
