+++ src/ec/packages/ec-component-site-header/demo/data--fr.js
@@ -1,0 +1,4 @@
+import specs from '@ecl/ec-specs-site-header/demo/data--fr';
+import adapter from '../adapter';
+
+export default adapter(specs);
