+++ src/ec/packages/ec-component-site-header/site-header.story.js
@@ -1,3 +1,4 @@
+import merge from 'deepmerge';
 import { storiesOf } from '@storybook/html';
 import { withKnobs } from '@storybook/addon-knobs';
 import { withNotes } from '@ecl-twig/storybook-addon-notes';
@@ -7,7 +8,8 @@ import defaultSprite from '@ecl/ec-resources-icons/dist/sprites/icons.svg';
 import englishBanner from '@ecl/ec-resources-logo/logo--en.svg';
 import frenchBanner from '@ecl/ec-resources-logo/logo--fr.svg';
 
-import { englishData, frenchData } from './demo/data';
+import englishData from './demo/data--en';
+import frenchData from './demo/data--fr';
 
 import siteHeader from './ecl-site-header.html.twig';
 import notes from './README.md';
@@ -18,22 +20,30 @@ storiesOf('Components/deprecated/Site Header', module)
   .addDecorator(withCode)
   .add(
     'ECL < 2.12 - default',
-    () => {
-      englishData.icon_file_path = defaultSprite;
-      englishData.logo.src = englishBanner;
-      return siteHeader(englishData);
-    },
+    () =>
+      siteHeader(
+        merge(englishData, {
+          logo: {
+            src: englishBanner,
+          },
+          icon_file_path: defaultSprite,
+        })
+      ),
     {
       notes: { markdown: notes, json: englishData },
     }
   )
   .add(
     'ECL < 2.12 - translated',
-    () => {
-      frenchData.icon_file_path = defaultSprite;
-      frenchData.logo.src = frenchBanner;
-      return siteHeader(frenchData);
-    },
+    () =>
+      siteHeader(
+        merge(frenchData, {
+          logo: {
+            src: frenchBanner,
+          },
+          icon_file_path: defaultSprite,
+        })
+      ),
     {
       notes: { markdown: notes, json: frenchData },
     }
